Miscellaneous
I renamed the data file and argument to more precisely describe the data
in question: monitors.

In preparation for adding a Helm chart, I added a Containerfile; why not
prepare for the post-Docker world?

The file descriptor is now adjusted only if necessary.

And I added support for defining the HTTP method to be used when
polling.